+++ CHANGES.md
@@ -16,6 +16,7 @@
 * Update html5lib 0.999 to 0.99999999/1.0b9 (46dae3d)
 * Update PNotify library 2.0.1 to 2.1.0
 * Update profilehooks 1.4 to 1.8.2.dev0 (ee3f1a8)
+* Update Requests library 2.7.0 (5d6d1bc) to 2.9.1 (a1c9b84)
 
 
 ### 0.11.0 (2016-01-10 22:30:00 UTC)